feat(packages): extract @flc/auth — deduplicate AuthContext across apps
AuthContext.tsx (325 lines) was byte-for-byte identical in the main app
and hrms-web. @flc/auth extracts the canonical implementation; apps
become thin wrappers that inject their local logger/errorTracker.
