+++ apps/hrms-web/tsconfig.json
@@ -17,6 +17,7 @@
       "@/hooks/use-mobile": ["../../packages/ui/src/hooks/use-mobile.tsx"],
       "@/*": ["../../src/*"],
       "@hrms-web/*": ["./src/*"],
+      "@flc/auth": ["../../packages/auth/src/index.ts"],
       "@flc/types": ["../../packages/types/src/index.ts"],
       "@flc/supabase": ["../../packages/supabase/src/index.ts"],
       "@flc/shell": ["../../packages/shell/src/index.ts"],
