+++ packages/auth/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@flc/auth",
+  "version": "0.0.1",
+  "private": true,
+  "description": "Shared authentication context, guards, and hooks for FLC BI apps",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "peerDependencies": {
+    "@supabase/supabase-js": "^2.0.0",
+    "react": "^18.0.0",
+    "react-router-dom": "^6.0.0"
+  }
+}
