+++ packages/auth/src/index.ts
@@ -1,0 +1,11 @@
+export {
+  AuthProvider,
+  useAuth,
+  ProtectedRoute,
+} from './AuthContext';
+export type {
+  AuthLogger,
+  AuthErrorTracker,
+  Profile,
+  AuthContextType,
+} from './AuthContext';
